Change agc enable to uint16_t

+++ lib/source_impl.cc
@@ -313,7 +313,7 @@ void source_impl::set_tcxo_dac(uint16_t dacVal) {
     device_handler::getInstance().set_tcxo_dac(stored.device_number, dacVal);
 }
 
-void source_impl::set_agc(int enabled) {
+void source_impl::set_agc(uint16_t enabled) {
     device_handler::getInstance().set_agc(stored.device_number, enabled);
 }
